@@ -1,37 +1,47 @@
 {
-  "name": "gamification",
-  "version": "1.0.0",
+  "name": "tests",
+  "version": "0.0.0",
+  "license": "MIT",
   "scripts": {
-    "prestart": "npm run build",
-    "start": "concurrently \"npm run build:watch\" \"npm run serve\" ",
-    "serve": "lite-server -c=bs-config.json",
-    "build": "tsc -p src/ ",
-    "build:watch": "tsc -p src/ -w"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
-
+  "private": true,
   "dependencies": {
-    "@angular/common": "~4.0.0",
-    "@angular/compiler": "~4.0.0",
-    "@angular/core": "~4.0.0",
-    "@angular/forms": "~4.0.0",
-    "@angular/http": "~4.0.0",
-    "@angular/platform-browser": "~4.0.0",
-    "@angular/platform-browser-dynamic": "~4.0.0",
-    "@angular/router": "~4.0.0",
-    "@types/lodash": "^4.14.54",
-
-    "angular-in-memory-web-api": "~0.3.0",
-    "systemjs": "0.19.40",
+    "@angular/common": "^4.0.0",
+    "@angular/compiler": "^4.0.0",
+    "@angular/core": "^4.0.0",
+    "@angular/forms": "^4.0.0",
+    "@angular/http": "^4.0.0",
+    "@angular/platform-browser": "^4.0.0",
+    "@angular/platform-browser-dynamic": "^4.0.0",
+    "@angular/router": "^4.0.0",
     "core-js": "^2.4.1",
-    "rxjs": "5.0.1",
-    "zone.js": "^0.8.4"
+    "rxjs": "^5.1.0",
+    "zone.js": "^0.8.4",
+    "angular-in-memory-web-api": "~0.3.0"
   },
-
   "devDependencies": {
-    "@types/node": "^6.0.46",
-    "concurrently": "^3.2.0",
-    "lite-server": "^2.2.2",
-    "typescript": "~2.1.0",
-    "lodash": "^4.16.4"
+    "@angular/cli": "1.0.2",
+    "@angular/compiler-cli": "^4.0.0",
+    "@types/jasmine": "2.5.38",
+    "@types/node": "~6.0.60",
+    "codelyzer": "~2.0.0",
+    "jasmine-core": "~2.5.2",
+    "jasmine-spec-reporter": "~3.2.0",
+    "karma": "~1.4.1",
+    "karma-chrome-launcher": "~2.0.0",
+    "karma-cli": "~1.0.1",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "protractor": "~5.1.0",
+    "ts-node": "~2.0.0",
+    "tslint": "~4.5.0",
+    "typescript": "~2.2.0"
   }
 }
