@@ -1,8 +1,10 @@
 import {Component} from '@angular/core';
 
 @Component({
-  selector: 'my-app',
-  templateUrl: 'app/app.component.html'
+  selector: 'app-gamification',
+  templateUrl: './app.component.html'
 })
-export class AppComponent{}
+
+export class AppComponent {
+}
 
