@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { AllUsersTableComponent } from './components/all-users-table/all-users-table.component';
-import {UserDetailsTableComponent} from "./components/user-details-table/user-details-table.component";
+import {UserDetailsTableComponent} from './components/user-details-table/user-details-table.component';
 
 const routes: Routes = [
     {
