@@ -2,7 +2,8 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-header',
-  templateUrl: 'app/components/header/header.component.html'
+  templateUrl: './header.component.html',
+   styleUrls: ['./header.component.css'],
 })
 
 export class HeaderComponent {
