@@ -2,5 +2,5 @@ export class User {
   uuid: string;
   name: string;
   skype: string;
-  slack: string
+  slack: string;
 }
