@@ -1,7 +1,7 @@
 import {Achievement} from './achievement';
 
 export class UserDetails {
-// uuid: string;
+  // uuid: string;
   // achievements: Achievement[];
   user: string;
   details: Achievement[];
